+++ src/components/Main/Main.test.tsx
@@ -1,12 +1,17 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
+import { BrowserRouter as Router } from 'react-router-dom'
 
 import Main from '.'
 
 describe('Main', () => {
   it('renders without crashing', () => {
     const div = document.createElement('div')
-    ReactDOM.render(<Main />, div)
+    ReactDOM.render(
+      <Router>
+        <Main />
+      </Router>,
+      div)
     ReactDOM.unmountComponentAtNode(div)
   })
 })